sample rate should be returned as uint64

+++ music.go
@@ -89,8 +89,8 @@ func (p *Player) Time() uint64 {
 	return t
 }
 
-func (p *Player) Rate() float64 {
-	return p.sampleRate
+func (p *Player) Rate() uint64 {
+	return uint64(p.sampleRate)
 }
 
 func (p *Player) process(data [][]float32) {
